Revive sparse matrix mul test + optimizations
Revive the sparse matrix multiply test which had a bug in it previously.
Add benchmarks for vector additions, which have the benchmarking
advantage of sidestepping any flatk overhead altogether.
This uncovered some optimization opportunities in the iteration code.
Some improvements are made with this commit (30% performance increase
for collecting results into a Vec). Further improvements need to be
done.